+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - [\#730](https://github.com/cosmos/evm/pull/730) Fix panic if evm mempool not used.
 - [\#733](https://github.com/cosmos/evm/pull/733) Avoid rejecting tx with unsupported extension option for ExtensionOptionDynamicFeeTx.
 - [\#736](https://github.com/cosmos/evm/pull/736) Add InitEvmCoinInfo upgrade to avoid panic when denom is not registered.
+- [\#732](https://github.com/cosmos/evm/pull/732) Fix gas meter race condition in integration tests
 
 ### IMPROVEMENTS